+++ CotEditor/Sources/Scripting Support/Document+ScriptingSupport.swift
@@ -213,7 +213,7 @@ extension Document {
         guard
             let arguments = command.evaluatedArguments,
             let encodingName = arguments["newEncoding"] as? String,
-            let encoding = EncodingManager.encoding(name: encodingName) ?? EncodingManager.encoding(ianaCharSetName: encodingName)
+            let encoding: String.Encoding = .init(localizedName: encodingName) ?? .init(ianaCharSetName: encodingName)
         else {
             command.scriptErrorNumber = OSAParameterMismatch
             command.scriptErrorString = "Invalid encoding name."
@@ -250,7 +250,7 @@ extension Document {
         guard
             let arguments = command.evaluatedArguments,
             let encodingName = arguments["newEncoding"] as? String,
-            let encoding = EncodingManager.encoding(name: encodingName) ?? EncodingManager.encoding(ianaCharSetName: encodingName)
+            let encoding: String.Encoding = .init(localizedName: encodingName) ?? .init(ianaCharSetName: encodingName)
         else {
             command.scriptErrorNumber = OSAParameterMismatch
             command.scriptErrorString = "Invalid encoding name."